chore(DTFS2-7545): revert updateDeal changes
