+++ dtfs-central-api/src/services/portal/deal.service.add-deal-cancelled-activity.test.ts
@@ -68,7 +68,7 @@ describe('PortalDealService - addGefDealCancelledActivity', () => {
   });
 
   describe(`when dealType is not ${DEAL_TYPE.GEF}`, () => {
-    it(`does not call addPortalActivity when dealType is not ${DEAL_TYPE.GEF}`, async () => {
+    it(`should not call addPortalActivity when dealType is not ${DEAL_TYPE.GEF}`, async () => {
       // Arrange
       const mockDeal = deal;
       mockDeal.dealSnapshot.dealType = DEAL_TYPE.BSS_EWCS;
