+++ dtfs-central-api/src/v1/controllers/portal/gef-deal/update-deal.controller.js
@@ -1,9 +1,40 @@
-const { validateAuditDetails } = require('@ukef/dtfs2-common/change-stream');
-const { InvalidAuditDetailsError } = require('@ukef/dtfs2-common');
+const { generateAuditDatabaseRecordFromAuditDetails, validateAuditDetails } = require('@ukef/dtfs2-common/change-stream');
+const { MONGO_DB_COLLECTIONS, InvalidAuditDetailsError } = require('@ukef/dtfs2-common');
 const { ObjectId } = require('mongodb');
 const { findOneDeal } = require('./get-gef-deal.controller');
+const { mongoDbClient: db } = require('../../../../drivers/db-client');
 const { isNumber } = require('../../../../helpers');
-const { updateDeal } = require('./update-deal');
+
+const updateDeal = async ({ dealId, dealUpdate, auditDetails }) => {
+  try {
+    if (!ObjectId.isValid(dealId)) {
+      return { status: 400, message: 'Invalid Deal Id' };
+    }
+
+    const collection = await db.getCollection(MONGO_DB_COLLECTIONS.DEALS);
+    const originalDeal = await findOneDeal(dealId);
+    const auditRecord = generateAuditDatabaseRecordFromAuditDetails(auditDetails);
+
+    const dealUpdateForDatabase = {
+      ...originalDeal,
+      ...dealUpdate,
+      updatedAt: Date.now(),
+      auditRecord,
+    };
+
+    const findAndUpdateResponse = await collection.findOneAndUpdate(
+      { _id: { $eq: ObjectId(String(dealId)) } },
+      { $set: dealUpdateForDatabase },
+      { returnNewDocument: true, returnDocument: 'after' },
+    );
+
+    return findAndUpdateResponse.value;
+  } catch (error) {
+    console.error('Unable to update deal %s %o', dealId, error);
+    return { status: 500, message: error };
+  }
+};
+exports.updateDeal = updateDeal;
 
 exports.updateDealPut = async (req, res) => {
   const {
